+++ src/Graph.js
@@ -8,38 +8,40 @@
 //
 //arguments :
 //  game            : reference to current game 
-//  x,y             : position of graph relative to top-left corner of game
-//  scale           : font-size and width of <pre> times this scalevalues(default 1)
-//  refreshRate     : draw graph every refreshRate ms
 //  maxY (in dev)   : the highest value of your input variable
+//  refreshRate     : draw graph every refreshRate ms
+//  bits            : retro mode, do not reformat 0's and 1's
 //  label (in dev)  : the label of your input variable
-//  input (in dev)  : your input variable
-//  bitMode         : retro mode, do not reformat 0's and 1's
 
-var Graph = function (game,x,y,scale,refreshRate,maxY,label,bitMode)
+var Graph = function (game,maxY,refreshRate,bits,label)
 {
-  this.game=game,this.scale=scale,this.label=label,this.maxY=maxY;this.bitMode=bitMode;
-  //explained in Graph.draw(), the zeros are used to pad the final string result.
-  this.scanBinary=this.shiftCount=0;this.zeros= Array(37).join("0");
-  // add the 5 scanlines      
-  this.line0=new Scanline(this,0),this.line1=new Scanline(this,1),
-  this.line2=new Scanline(this,2),this.line3=new Scanline(this,3),
-  this.line4=new Scanline(this,4);
-  return this;
+    this.game=game;this.refreshRate=this.initialRate=refreshRate;this.maxY=maxY;this.label=label;this.bits=bits;this.hide=false;
+    //explained in Graph.draw(), the zeros are used to pad the final string result.
+    this.scanBinary=this.startBinary=0x010000000;this.zeros= Array(30).join("0"); this.counter=0;
+    // add the scanlines      
+    this.line0=new Scanline(this,0);this.line1=new Scanline(this,1);
+    this.line2=new Scanline(this,2);this.line3=new Scanline(this,3);
+    this.line4=new Scanline(this,4);
+    return this;
 }
-//graph drawing, concatenates Scanline.draw() outputs, then displays the string
-Graph.prototype.draw =function()
-{
-  //which input to draw
-  this.input=this.label=="FPS"?this.game.time.fps:this.game.time.elapsedMS;
-  //60 fps->row 0, 50 fps->row 1 , 40 fps->row 2.......
-  this.rownumber=~~((this.input)/(this.maxY/5));
-  //scanbinary is a reference used for masking).if reached MSB, reset to zero, else shift by one bit
-  this.scanBinary=1>>this.scanBinary==0?0:1<<( this.shiftCount);
-  //accumulator, used for scanBinary (see comment at 91); 
-  this.shiftCount=this.shiftCount-1;
-  //concatenate and display resulting string
-  this.result=this.line4.draw()+this.line3.draw()+this.line2.draw()+this.line1.draw()+this.line0.draw();
-  this.result+=this.input+" "+this.label+ " DC: "+this.game.renderer.renderSession.drawCount;
-  this.node.textContent=this.result;
+
+//graph drawing, if not hidden, concatenate output of Scanline.draw(), then returns the string
+
+Graph.prototype.update =function(input)
+{  
+    //clock
+    this.counter=(this.counter+1)%this.refreshRate;
+    if(!this.hide&&this.counter==0)
+    {
+        //determine which scanline to draw on.
+        this.input=input<1?input*1000:input;
+        this.rownumber=~~((this.input-1)/(this.maxY/5));
+        //scanbinary is a reference used for masking.if reached MSB, start over, else shift by one bit
+        this.scanBinary=((this.scanBinary>>1))||this.startBinary;
+        //concatenate and display resulting string
+        this.result=this.line4.draw()+this.maxY+'\n'+this.line3.draw()+'\n'+this.line2.draw()+'\n'+this.line1.draw()+'\n'+this.line0.draw()+'\n';
+        this.result+=this.input+" "+this.label+" DC:"+this.game.renderer.renderSession.drawCount;
+    }
+    return this.result;
 };
+