+++ CHANGELOG.md
@@ -10,6 +10,56 @@
 
 ### Thanks
 
+## Version 1.0.6 - 9th September 2017
+
+### Updates
+
+* Switch colors by tapping graph.
+* Cycle trough graphs with the new panel layout.
+* Added UI buttons to: toggle, resize, speedup, cycle.
+* Added detail mode to measure state,stage,plugins,tweens,sound,input,physics,particles updates.
+
+### Bug Fixes
+
+### Documentation
+
+### Thanks
+
+@samid737
+
+## Version 1.0.5 - 5th September 2017
+
+### Updates
+
+* added color change via hue shift
+* added graph visibility toggling
+* reverted to 1.02
+* bitmode re-added, for supreme retro feel.
+
+### Bug Fixes
+
+### Documentation
+
+### Thanks
+
+@samid737
+
+## Version 1.0.3 - 4th September 2017
+
+### Updates
+
+* removed FPS update 
+* replaced string formatting (0 to _ and 1 to *) with custom font. up to 50% decrease in execution time
+* removed bitmode
+
+### Bug Fixes
+
+### Documentation
+
+### Thanks
+
+@samid737
+
 ## Version 1.0.2 - 4th September 2017
 
 ### Updates